Updated toString methods of most classes that will print state of all its members. Added new class called ClassStateString to facilitate the formatting of the toString method. Classes no longer use have the option to draw directly on a bitmap. Added Callbacks that log which state the activity is in. Added Button that when pressed, the current state is logged. Removed SwipeGestureDetector sine it has been replaced with MultiSwipeListener.
Activity is now implements onRestoreInstanceState and onSaveInstanceState.
GameSurfaceView now implements onSaveInstanceState and onRestoreInstanceState.
Classes with important information now have a function to save and restore its state from a bundle.
Added new String members to classes to be used as keys when saving and restoring state.

fixed a bug where GameSurfaceView would Cycle data was lost when the Surface was destroyed.