Began preparations to make surface view show text instead of separate image views.
This is being done because SurfaceView is always on top of everything and TextViews
would be hidden underneath it.

plan a:  find a way to find a way to make a view appear over it.
plan b:  make the GameSurfaceView display the text
plan c: make the GameSurfaceView draw the background always.

attempting plan b

MultiplayerActivity
     - removed go text prompt
     - added 5 private variables that hold text that will be displayed.
     - removed mCountdownAlarmID since its being functionality is being into
        mReadyCountdownAlarmId and mSetCountdownAlarmId.
     - removed mSetCountdownAlarmId since go is being removed as a text prompt.
     - removed mCountDownPrompt since it did not handle click event and no longer needs to show text.

GameSurfaceView
     - constructor now initializes a paint private member, this will be used to writing text
     - setText writes text onto the view
     - removeText removes any text from the view
     - redrawView now draws text

multiplayer_game_activity_xml
     - removed countdown textView